Remove verbose and move constants to metadata
